Extract find filename at cursor to VimSearchHelper
